+++ tests/unit_tests/drawtypes/ramp.cpp
@@ -1,5 +1,6 @@
-#include "common/test.hpp"
 #include "drawtypes/ramp.hpp"
+
+#include "common/test.hpp"
 #include "utils/factory.hpp"
 
 using namespace polybar::drawtypes;
@@ -22,4 +23,27 @@ TEST(Ramp, perc) {
   r.add(factory_util::shared<label>("test4", 0));
   EXPECT_EQ("test2", r.get_by_percentage_with_borders(29, 20, 40)->get());
   EXPECT_EQ("test3", r.get_by_percentage_with_borders(31, 20, 40)->get());
+}
+
+TEST(Ramp, weights) {
+  ramp r;
+  r.add(factory_util::shared<label>("test1", 0), 1);
+  r.add(factory_util::shared<label>("test2", 0), 2);
+  r.add(factory_util::shared<label>("test3", 0), 5);
+
+  EXPECT_EQ("test1", r.get_by_percentage(12)->get());
+  EXPECT_EQ("test2", r.get_by_percentage(13)->get());
+  EXPECT_EQ("test2", r.get_by_percentage(37)->get());
+  EXPECT_EQ("test3", r.get_by_percentage(38)->get());
+
+  EXPECT_EQ("test1", r.get_by_percentage_with_borders(19, 20, 40)->get());
+  EXPECT_EQ("test2", r.get_by_percentage_with_borders(21, 20, 40)->get());
+  EXPECT_EQ("test3", r.get_by_percentage_with_borders(39, 20, 40)->get());
+  EXPECT_EQ("test3", r.get_by_percentage_with_borders(41, 20, 40)->get());
+  EXPECT_EQ("test1", r.get_by_percentage_with_borders(20, 20, 40)->get());
+  EXPECT_EQ("test3", r.get_by_percentage_with_borders(40, 20, 40)->get());
+  r.add(factory_util::shared<label>("test4", 0));
+  r.add(factory_util::shared<label>("test5", 0));
+  EXPECT_EQ("test2", r.get_by_percentage_with_borders(24, 20, 40)->get());
+  EXPECT_EQ("test3", r.get_by_percentage_with_borders(25, 20, 40)->get());
 }