+++ CHANGELOG.md
@@ -52,6 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `POLYBAR_FLAGS` cmake variable can be used to pass extra C++ compiler flags.
 
 ### Added
+- `drawtypes/ramp`: Add support for ramp weights.
+   ([1750](https://github.com/polybar/polybar/issues/1750))
 - `internal/memory`: New tokens `%used%`, `%free%`, `%total%`, `%swap_total%`, 
   `%swap_free%`, and `%swap_used%` that automatically switch between MiB and GiB
   when below or above 1GiB.